# backend/MeetMyBarApi/src/main/resources/swagger/meetMyBarSwagger.yaml
openapi: 3.1.0
...
info:
  title: meetMyBarSwagger
  version: '1.0'
  summary: Meet my bar
  description: "Meet My Bar est une application mobile Android qui permet de repérer les bars autour de soi et d'avoir des informations pratiques comme le prix de la bière la moins chère. Il sera possible de rechercher un bar par son nom, par un type de bière ou par tranche de prix pour les bières. Il est possible de se connecter  et de remplir ses infos personnelles et d’ajouter des amis sur l’application. Ajouter des amis permet d’avoir accès à leur favori et d’être ajouté à des événements créés pour une sortie dans un des bars présents. Les bars possèdent des évènements où l’utilisateur peut choisir si l'événement est visible par tous ou seulement pas ses invités ( au sein de ses amis ). Un événement est propre à un bar et à une date et heure précise. L’application est renseignée par ses utilisateurs, chaque user peut créer un bar et renseigner les informations qui lui correspondent, il est ensuite disponible pour les autres utilisateurs.\r\nIl peut également mettre à jour les données du bar avec le nouveaux prix de la bière, ses horaires d’ouverture/fermeture.\r\n"
  contact:
    name: ''
servers:
  - url: 'http://localhost:8080'
    description: ''
paths:
...
  /drink:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drink'
                  x-stoplight:
                    id: 1hhzgtmz8glt7
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: get-drink
      x-stoplight:
        id: azpwkc6to7yjk
    post:
      summary: Your POST endpoint
      tags: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: post-drink
      x-stoplight:
        id: 3xhugknok8mz9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drink'
    patch:
      summary: Your PATCH endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: patch-drink
      x-stoplight:
        id: 6fcbfsf1u8fds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drink'
    delete:
      summary: Your DELETE endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: delete-drink
      x-stoplight:
        id: 7fmq3m5xwnlrt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drink'
  '/drink/{drinkName}':
    parameters:
      - schema:
          type: string
        name: drinkName
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: get-drink.-drinkName
      x-stoplight:
        id: 6tyqiqc517buq
components:
  schemas:
...
    Drink:
      title: Drink
      x-stoplight:
        id: sgpqj4r6lfb75
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          x-stoplight:
            id: 0av3dcnadccfo
        alcohol_degree:
          type: number
          x-stoplight:
            id: wbg1e7uw78w95
        name:
          type: string
          x-stoplight:
            id: hj2pc69610sz3
        brand:
          type: string
          x-stoplight:
            id: drvto8xezd3ea
      x-tags:
        - drink
...